Remove sub package “commands” and simply put everything in main package, but still distributed across multiple files. Updated ./bin/run to work with multiple main package files.
