+++ create.go
@@ -1,9 +1,9 @@
-package commands
+package main
 
-import "fmt"
 import "os"
+import "fmt"
 
-func Destroy(args []string) {
+func Create(args []string) {
 	if len(args) == 0 {
 		fmt.Fprintln(os.Stderr, "Missing argument: <name>")
 		os.Exit(1)
@@ -17,18 +17,18 @@ func Destroy(args []string) {
 		os.Exit(1)
 	}
 
-	if !directoryExists(client, name) {
-		fmt.Fprintln(os.Stderr, "Repository", name, "doesn't exist.")
+	if repositoryExists(client, name) {
+		fmt.Fprintln(os.Stderr, "Repository", name, "already exists.")
 		os.Exit(1)
 	}
 
-	command := "rm -rf " + name + ".git"
+	command := "mkdir -p " + name + ".git && cd " + name + ".git && git --bare init"
 	_, stderr, err := client.Run(command)
 	if err != nil {
-		fmt.Fprintln(os.Stderr, "Failed to destroy repository.")
+		fmt.Fprintln(os.Stderr, "Failed to create repository.")
 		fmt.Fprintln(os.Stderr, err.Error(), stderr)
 		os.Exit(1)
 	}
 
-	fmt.Println("Repository", name, "destroyed.")
+	fmt.Println("Repository", name, "created.")
 }
