+++ helpers_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import "os"
 import "testing"
