+++ list_test.go
@@ -1,4 +1,4 @@
-package commands
+package main
 
 import "testing"
 import "github.com/stretchr/testify/assert"
